+++ src/Admin.php
@@ -7,8 +7,12 @@ class Admin {
 
 	public Settings $settings;
 
-	public function __construct( Settings $settings ) {
-		$this->settings = $settings;
+	public $plugin_url;
+
+
+	public function __construct( Settings $settings, string $plugin_url ) {
+		$this->settings   = $settings;
+		$this->plugin_url = $plugin_url;
 
 		if ( ! $this->settings->get_openai_api_key() ) {
 			return;
@@ -20,6 +24,10 @@ class Admin {
 
 		if ( $this->settings->should_generate_alt_text() ) {
 			add_filter( 'wp_update_attachment_metadata', [ $this, 'update_alt_text' ], 10, 2 );
+		}
+
+		if ( $this->settings->should_integrate_dall_e() ) {
+			add_action( 'enqueue_block_editor_assets', $this->enqueue_scripts( ... ) );
 		}
 	}
 
@@ -77,5 +85,22 @@ class Admin {
 		}
 
 		return $data;
+	}
+
+	public function enqueue_scripts() {
+		$version_file = __DIR__ . '/../build/index.build.asset.php';
+		if ( ! file_exists( $version_file ) ) {
+			return;
+		}
+		$version = include $version_file;
+		wp_enqueue_script(
+			'better-file-name-ai',
+			$this->plugin_url . '/index.build.js',
+			$version['dependencies'],
+			$version['version'],
+			[
+				'in_footer' => true,
+			]
+		);
 	}
 }
